requirements.txt
streamlit 
langchain_groq
langchain_community 
PyPDF
pypdf2
google.generativeai
python-dotenv
-e . 